scraper/requirements.txt
boto3
protobuf==4.21.6
google-api-python-client
tqdm
google-generativeai
